2005-03-17  Erik Hjortsberg  <erik@katastrof.nu>
	* AvatarEmberEntity.h: better looking comments
	* TerrainGenerator.cpp: cast double to int

+++ src/components/ogre/terrain/TerrainGenerator.cpp
@@ -203,10 +203,10 @@ void TerrainGenerator::prepareAllSegments(bool alsoPushOntoTerrain)
 	}
 
 	int mercatorSegmentsPerPage =  getSegmentSize() / 64;
-	int xNumberOfPages = ceil((mXmax - mXmin) / (double)mercatorSegmentsPerPage);
-	int yNumberOfPages = ceil((mYmax - mYmin) / (double)mercatorSegmentsPerPage);
-	int xStart = floor(mXmin / (double)(mercatorSegmentsPerPage));
-	int yStart = floor(mYmin / (double)(mercatorSegmentsPerPage));
+	int xNumberOfPages = (int)ceil((mXmax - mXmin) / (double)mercatorSegmentsPerPage);
+	int yNumberOfPages = (int)ceil((mYmax - mYmin) / (double)mercatorSegmentsPerPage);
+	int xStart = (int)floor(mXmin / (double)(mercatorSegmentsPerPage));
+	int yStart = (int)floor(mYmin / (double)(mercatorSegmentsPerPage));
 	for (i = 0; i < xNumberOfPages; ++i) {
 		for (j = 0; j < yNumberOfPages; ++j) {
 			TerrainPosition pos(xStart + i, yStart + j);
